Introduce ExportFactory
